Fix Win64 build.
MemoryAllocator::AvailableExecutable should return intptr_t, the difference of intptr_t's.

TBR=ager@chromium.org

Review URL: http://codereview.chromium.org/4972001

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@5818 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/spaces.h
@@ -591,7 +591,7 @@ class MemoryAllocator : public AllStatic {
   static intptr_t Size() { return size_; }
 
   // Returns the maximum available executable bytes of heaps.
-  static int AvailableExecutable() {
+  static intptr_t AvailableExecutable() {
     if (capacity_executable_ < size_executable_) return 0;
     return capacity_executable_ - size_executable_;
   }
